Create a class for rendering the bar chart
