+++ src/org/missionassetfund/apps/android/adapters/TransactionsExpandableListAdapter.java
@@ -3,7 +3,6 @@ package org.missionassetfund.apps.android.adapters;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
