+++ src/org/missionassetfund/apps/android/fragments/DailyTransactionsChartFragment.java
@@ -6,24 +6,19 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import org.achartengine.ChartFactory;
 import org.achartengine.GraphicalView;
 import org.achartengine.chart.BarChart.Type;
-import org.achartengine.model.CategorySeries;
 import org.achartengine.model.SeriesSelection;
-import org.achartengine.model.XYMultipleSeriesDataset;
-import org.achartengine.renderer.SimpleSeriesRenderer;
-import org.achartengine.renderer.XYMultipleSeriesRenderer;
 import org.missionassetfund.apps.android.R;
 import org.missionassetfund.apps.android.models.Category;
 import org.missionassetfund.apps.android.models.CategoryTotal;
 import org.missionassetfund.apps.android.models.Chart;
 import org.missionassetfund.apps.android.models.dao.CategoryDao;
+import org.missionassetfund.apps.android.views.BarChart;
 
 import android.app.Activity;
 import android.content.Context;
 import android.graphics.Color;
-import android.graphics.Paint.Align;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
@@ -132,71 +127,12 @@ public class DailyTransactionsChartFragment extends Fragment {
 
         String[] xTitles = xLabels.toArray(new String[xLabels.size()]);
 
-        XYMultipleSeriesRenderer renderer = buildBarRenderer(categoriesColors);
-        setChartSettings(renderer, X_VALUES_EDGE,
-                MAX_CHART_VALUES + X_VALUES_EDGE, 0, maxValue.floatValue(), Color.GRAY,
-                Color.LTGRAY, xTitles);
-
-        mGraphicalView = ChartFactory.getBarChartView(context,
-                buildBarDataset(categoriesTitles, values), renderer,
-                Type.STACKED);
+        BarChart barChart = new BarChart(categoriesColors, X_VALUES_EDGE, MAX_CHART_VALUES, maxValue.floatValue(), xTitles);
+        mGraphicalView = barChart.getChartView(context, categoriesTitles, values, Type.STACKED); 
 
         mGraphicalView.setOnClickListener(chartClickListener);
 
         rlStackedBarChart.addView(mGraphicalView);
-    }
-
-    protected XYMultipleSeriesRenderer buildBarRenderer(int[] colors) {
-        XYMultipleSeriesRenderer renderer = new XYMultipleSeriesRenderer();
-        int length = colors.length;
-        for (int i = 0; i < length; i++) {
-            SimpleSeriesRenderer r = new SimpleSeriesRenderer();
-            r.setColor(colors[i]);
-            renderer.addSeriesRenderer(r);
-        }
-        return renderer;
-    }
-
-    protected void setChartSettings(XYMultipleSeriesRenderer renderer, double xMin, double xMax,
-            double yMin, double yMax, int axesColor,
-            int labelsColor, String[] xTitles) {
-        renderer.setXAxisMin(xMin);
-        renderer.setXAxisMax(xMax);
-        renderer.setYAxisMin(yMin);
-        renderer.setYAxisMax(yMax);
-        renderer.setAxesColor(axesColor);
-        renderer.setLabelsColor(labelsColor);
-        renderer.setMarginsColor(Color.WHITE);
-        renderer.setShowLegend(false);
-        renderer.setXLabels(0);
-
-        for (int i = 0; i < xTitles.length; i++) {
-            renderer.addXTextLabel(i + 1, xTitles[i]);
-        }
-
-        renderer.setXLabelsAlign(Align.CENTER);
-        renderer.setPanEnabled(false, false);
-        renderer.setZoomRate(1.1f);
-        renderer.setBarSpacing(0.5f);
-        renderer.setLabelsTextSize(24);
-        renderer.setYLabels(0);
-        renderer.setShowCustomTextGrid(true);
-        renderer.setClickEnabled(true);
-    }
-
-    protected XYMultipleSeriesDataset buildBarDataset(String[] titles, List<double[]> values) {
-        XYMultipleSeriesDataset dataset = new XYMultipleSeriesDataset();
-        int length = titles.length;
-        for (int i = 0; i < length; i++) {
-            CategorySeries series = new CategorySeries(titles[i]);
-            double[] v = values.get(i);
-            int seriesLength = v.length;
-            for (int k = 0; k < seriesLength; k++) {
-                series.add(v[k]);
-            }
-            dataset.addSeries(series.toXYSeries());
-        }
-        return dataset;
     }
 
     public void setChart(Chart chart) {
